Implement suggested minor refactoring & javadoc changes
